Move profile files to profile dir
