+++ configure.ac
@@ -281,7 +281,7 @@ AC_CONFIG_LINKS([include/souffle/CompiledOptions.h:src/CompiledOptions.h])
 AC_CONFIG_LINKS([include/souffle/RecordTable.h:src/RecordTable.h])
 AC_CONFIG_LINKS([include/souffle/CompiledSouffle.h:src/CompiledSouffle.h])
 AC_CONFIG_LINKS([include/souffle/CompiledTuple.h:src/CompiledTuple.h])
-AC_CONFIG_LINKS([include/souffle/EventProcessor.h:src/EventProcessor.h])
+AC_CONFIG_LINKS([include/souffle/EventProcessor.h:src/profile/EventProcessor.h])
 AC_CONFIG_LINKS([include/souffle/Explain.h:src/Explain.h])
 AC_CONFIG_LINKS([include/souffle/ExplainProvenance.h:src/ExplainProvenance.h])
 AC_CONFIG_LINKS([include/souffle/ExplainProvenanceImpl.h:src/ExplainProvenanceImpl.h])
@@ -289,10 +289,10 @@ AC_CONFIG_LINKS([include/souffle/ExplainTree.h:src/ExplainTree.h])
 AC_CONFIG_LINKS([include/souffle/EquivalenceRelation.h:src/EquivalenceRelation.h])
 AC_CONFIG_LINKS([include/souffle/IOSystem.h:src/io/IOSystem.h])
 AC_CONFIG_LINKS([include/souffle/LambdaBTree.h:src/LambdaBTree.h])
-AC_CONFIG_LINKS([include/souffle/Logger.h:src/Logger.h])
+AC_CONFIG_LINKS([include/souffle/Logger.h:src/profile/Logger.h])
 AC_CONFIG_LINKS([include/souffle/PiggyList.h:src/PiggyList.h])
-AC_CONFIG_LINKS([include/souffle/ProfileDatabase.h:src/ProfileDatabase.h])
-AC_CONFIG_LINKS([include/souffle/ProfileEvent.h:src/ProfileEvent.h])
+AC_CONFIG_LINKS([include/souffle/ProfileDatabase.h:src/profile/ProfileDatabase.h])
+AC_CONFIG_LINKS([include/souffle/ProfileEvent.h:src/profile/ProfileEvent.h])
 AC_CONFIG_LINKS([include/souffle/RamTypes.h:src/RamTypes.h])
 AC_CONFIG_LINKS([include/souffle/ReadStream.h:src/io/ReadStream.h])
 AC_CONFIG_LINKS([include/souffle/ReadStreamCSV.h:src/io/ReadStreamCSV.h])
