+++ src/TypeAttribute.cpp
