+++ src/interpreter/InterpreterEngine.cpp
@@ -17,8 +17,8 @@
 #include "AggregateOp.h"
 #include "BinaryConstraintOps.h"
 #include "FunctorOps.h"
-#include "Logger.h"
-#include "ProfileEvent.h"
+#include "profile/Logger.h"
+#include "profile/ProfileEvent.h"
 #include "RamTypes.h"
 #include "RecordTable.h"
 #include "SignalHandler.h"
