+++ src/profile/Reader.h
@@ -8,8 +8,8 @@
 
 #pragma once
 
-#include "../ProfileDatabase.h"
-#include "../ProfileEvent.h"
+#include "ProfileDatabase.h"
+#include "ProfileEvent.h"
 #include "Iteration.h"
 #include "ProgramRun.h"
 #include "Relation.h"
